+++ spec/lib/statistrano/deployment/strategy/releases_spec.rb
@@ -73,8 +73,8 @@ describe Statistrano::Deployment::Strategy::Releases do
 
       config_double = instance_double("Statistrano::Config")
       expect( Statistrano::Config ).to receive(:new)
-                                   .with(default_options.merge({remotes: [{hostname: 'web01', remote_dir: 'web01_remote_dir'}]})
-                                                        .merge(hostname: 'web01', remote_dir: 'web01_remote_dir'))
+                                   .with( options: default_options.merge({remotes: [{hostname: 'web01', remote_dir: 'web01_remote_dir'}]})
+                                                                  .merge(hostname: 'web01', remote_dir: 'web01_remote_dir'))
                                    .and_return(config_double)
       expect( Statistrano::Remote ).to receive(:new)
                                    .with(config_double)