+++ changelog.md
@@ -1,3 +1,4 @@
+- change the internal method signature of Config to use keyword arguments `options`, and `tasks`
 - add a `run_local` method to Remote to wrap `Shell.run_local` to display commands when in verbose mode.
 
 # 1.2.3